+++ app/javascript/components/AddPopup/AddPopup.js
@@ -36,7 +36,7 @@ function AddPopup({ onClose, onCreateCard }) {
       setSaving(false);
       setErrors(error || {});
 
-      setMessage({ type: SEVERITY.ERROR, text: `Task saving failed! ${error?.message || ''}` });
+      setMessage({ type: SEVERITY.ERROR, text: 'Task saving failed!' });
       setIsOpenSnackbar(true);
     });
   };
