+++ app/javascript/hooks/store/useTasks.js
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 
-import useTasksActions from 'slices/useTasksActions';
+import { useTasksActions } from 'slices/TasksSlice';
 import { COLUMNS } from 'constants/board';
 
 const useTasks = () => {
